Move App component to container folder, update import statement in index.js file
